Fix variance of choice key type in ledger.(createAnd)exercise
Fixes #8338

This is a bit of a weird one. The basic issue is fairly clear. We have
code which looks something like:

```
interface MyTemplate<K> {
  keyEncode: (k: K) => unknown;
  keyDecoder: () => K;
}

interface MyChoice<K = unknown> {
  template: () => MyTemplate<K>;
}

function exercise(c: MyChoice) {
  console.log(c);
}

const y = (x : MyChoice<undefined>) => {
  exercise(x);
}
```

The key type is invariant here since it occurs in both covariant and
contravariant positions. We compile in strict mode so this is a
typescript error. This fails in all typescript versions I tested
including 3.8 which we use in our tests. So why don’t the tests fail?

Afaict, this is because the template constant is defined as an
intersection of the template type and the choices available on the
template. In TypeScript 3.8, this seems to (mostly, we’ll get to a
bit) hide this error. In TypeScript 3.9 and later, you hit it. There
is something in the release notes about intersection types but only
about optional properties which don’t seem to be required
here. Nevertheless, it is definitely related to the typescript
version.

So there are two more weird things about TypeScript 3.8:

1. Sometimes you get the error in the IDE but not in `npm start` and
   `npm run build`. Afaict, this is caused by the VSCode plugin being
   based on a different TypeScript version, e.g., mine was at 4.1.2 so
   this seems to match up with the 3.9 hypothesis.

2. In some cases (not all afaict), the issue shows up in `npm start`
   but only intermittendly. As described in #8338, it goes away after
   a restart. I have no idea what is going on here. something
   something incremental builds are hard :shrug: I’ve spent way too
   much time trying to figure it out but since the underlying issue
   seems clear, I gave up at some point.

changelog_begin

- [JS Client Libraries] Fix a bug where in some cases calls to
  `exercise` and `createAndExercise` failed to typecheck with a mismatch
  in the key type.

changelog_end