faster cleanup on Linux
I have seen the cleanup step taking up to an hour, just to delete the
Bazel cache. This is obviously not great.

I've been researching how to do this faster, and it looks like the best
approach here is to create a partition for the cache, and just drop it
and reformat it when we want to clear it. This is a conceptually simple
approach, but it unfortunately requires access to the `mount` command.
There are a couple ways to achieve that, but after a bit of research it
seems to me that the safest one is to have a setuid-root executable that
normal users can call.

So this PR includes a new C program that's essentially a _safer_ version
of a bash script that would do the same thing, somehow, because it has
to run as root.

Please review carefully.

CHANGELOG_BEGIN
CHANGELOG_END