Enable ledger-api-test-tool to be used as a library
This is the third of four PRs in which https://github.com/digital-asset/daml/commit/6ea70c4b45d55cbf62caa3cc03dd70310ca79b87
has been broken up to facilitate review.

The endgame is to have the non-repudiation prototype merged. The
Ledger API test tool will be used to check that the proxy can be
safely put in front of a conformant ledger implementation without
causing conformance tests to not pass. The ability to use is as
a library facilitates the process, ensuring we don't have to set
up three running processes in Bazel (the participant, the proxy
and the test tool).

changelog_begin
changelog_end