participant-integration-api: Make the RecoveringIndexer use a real execution context. (#8688)
* participant-integration-api: Fix a RecoveringIndexerSpec flake.

Sometimes the log lines don't come in quite the same order, and that's
OK.

* participant-integration-api: Use the Akka EC to recover the indexer.

Using the DirectExecutionContext opens us up to abusing a random EC,
sometimes with big delays. Using the Akka EC is not quite ideal, but
still means that we know that we're not hogging an entire pool by
accident.

CHANGELOG_BEGIN
CHANGELOG_END

* participant-integration-api: Fix the RecoveringIndexerSpec for good.

We can fix it by filtering out the floating log line, instead of
guessing where it will end up.