ledger-api-test-tool: Increase the dedup time in CommandDeduplicationIT. (#8765)
The deduplication time was hard-coded to 3 seconds, which can be too
short for some ledgers; the command may have expired by the time it gets
to the commit stage.

In order to avoid spurious errors, this multiplies the deduplication
time by the timeout scale factor, allowing the ledger operator to
control the deduplication timeouts in the same way as all test timeouts.

CHANGELOG_BEGIN
- [Ledger API Test Tool] The CommandDeduplicationIT now sets the
  deduplication timeout on commands using the parameter passed to
  `--timeout-scale-factor`, allowing it to run on ledger configurations
  that require more than 3 seconds to commit a submission.
CHANGELOG_END