Add BigDecimal to the LF spec
There is one significant main difference compared to
https://github.com/digital-asset/daml/issues/8719:

- I removed `setScale` and more importantly `getScale`.

The
[Java spec](https://docs.oracle.com/javase/8/docs/api/java/math/BigDecimal.html)
provides the reasoning here. If you look at the docs around preferred
scale, this is quite tricky to specify and even the Java spec is
fairly vague. By keeping the scale abstract, we can avoid this whole
specification mess which seems nice.

That said, I could definitely be convinced that exposing the Scala is
sufficiently important that it is worth the complexity.

changelog_begin
changelog_end