1.8.x special patch fix transaction stream (#8971)
* Fix rare out of order transaction delivery on the ledger api (#8336)

* Fix out of order transaction delivery on the ledger api

Akka's mergeSubstreams does not guarantee the order of the output.
In rare cases the order of the output doesn't match with the order
of the input, even though we know that the input events are in the
correct order (otherwise groupContiguous would delivery transactions
with random events from other transactions).

This change removes the need for substreams by collecting the events
for a transaction inside statefulMapConcat. The loss of mergeSubstreams'
parallelism is likely no problem, because we anyway need to send out transactions
in the correct order.

Fixes #7521.

CHANGELOG_BEGIN
Ledger API: Fixed an issue that rarely caused transactions to be sent out of order. See #7521
CHANGELOG_END

* Use mergeSubstreamsWithParallelism(1)

* Use concatSubstreams

* Add test that can trigger the error reliably

* Address https://github.com/digital-asset/daml/pull/8336#discussion_r545760676

* Update copyright notice in GroupContiguousHeavySpec.scala

* Remove link to PR

Co-authored-by: Stefano Baghino <stefano.baghino@digitalasset.com>
Co-authored-by: Robert Autenrieth <robert.autenrieth@digitalasset.com>

* Add buffer to speed up transaction streaming (#8579)

* Add buffer to speed up transaction streaming

CHANGELOG_BEGIN
- [Ledger API Server] Fixed an issue that was causing a slowdown of the
  transaction service.
CHANGELOG_END

* Move the buffer to TransactionReader

Co-authored-by: Gerolf Seitz <gerolf.seitz@daml.com>
Co-authored-by: Stefano Baghino <stefano.baghino@digitalasset.com>
Co-authored-by: Robert Autenrieth <robert.autenrieth@digitalasset.com>
Co-authored-by: Robert Autenrieth <31539813+rautenrieth-da@users.noreply.github.com>