Fix protobuf structure for empty import prefix
https://github.com/digital-asset/daml/commit/052f69cde93d98eaeefe8938f5d33d2e80ca96eb
broke the structure of our protobufs by changing the import prefix
from "." to an empty string which ends up flattening the whole
file. This PR changes the default for tars to match the old
behavior. We can’t just change the default at the function level since
the prefix is used outside of pkg_tar and in those places an empty
string is required :(

changelog_begin
changelog_end