env vars
```
$ DAML_SDK_ADDRESS=0.0.0.0 daml-head json-api --ledger-host [...]
[...] - Config([...], address=0.0.0.0, [...])
```

My goal is to find a way to let users set options with environment
variables instead of CLI arguments. This is definitely not pretty, and
only touches the JSON API for now, but it seems to work.

I'm opening a PR to gather feedback on:

- Is this something we want? Is it worth spending more time on it?
- Is the approach completely wrong, and if so, what would you suggest
  instead?

One nice property of the code as it stands is that it uses the CLI arg
definition as the single source of truth, so there is no risk of CLI
args and env vars getting out of sync, and documentation could explain
the correspondance in a couple sentences. This could be reused across all
of our projects that use scopt.

Not so nice things:

- `scopt` is clearly not designed for this; this code, while simple and
  reasonably readable in my opinion, is definitely very hackish and may
  break when changing `scopt` version.
- At the moment required arguments have to be passed through the CLI;
  changing that is possible but a it more invasive. Given that we don't
  have that many required arguments, I think there is a case for not
  addressing that limitation if we want to keeo the code simple.

CHANGELOG_BEGIN
not yet
CHANGELOG_END