step 2: change one group
In this case, we know the Hoogle service works with just two machines,
so we don't need to scale up the green group. Similarly, the blue group
is already set to just one machine so we don't need to scale it down
before proceeding.

This step does require a `terraform apply`.