step 3: monitor the new deployment
In this case, this is mostly just looking at the logs of the blue
machine. I did it with

```
gcloud beta \
  compute instances \
  --project "da-dev-gcp-daml-language" \
  tail-serial-port-output \
  --zone "us-east4-a" \
  hoogle-blue-jcw2
```

Because it was so much faster, I did take a minute to log into the
machine and check everything was fine. In order to do that I had to
temporaily open the network, and I though I should also leave that code
in place to simplify future operations.